[buteo-sync-plugin-caldav] Add a tests.xml for test bot. Contributes …

+++ tests/notebooksyncagent/tst_notebooksyncagent.cpp
@@ -93,9 +93,10 @@ tst_NotebookSyncAgent::~tst_NotebookSyncAgent()
 
 void tst_NotebookSyncAgent::initTestCase()
 {
-    qputenv("SQLITESTORAGEDB", "./db");
-
-    QFile::remove("./db");
+    if (qgetenv("SQLITESTORAGEDB").isEmpty()) {
+        qputenv("SQLITESTORAGEDB", "./db");
+        QFile::remove("./db");
+    }
 }
 
 void tst_NotebookSyncAgent::cleanupTestCase()
